replace all read permissions by write permissions
