+++ app/src/main/java/com/infomaniak/drive/utils/Extensions.kt
@@ -531,54 +531,6 @@ fun Context.requestCredentials(onSuccess: () -> Unit) {
     }
 }
 
-@SuppressLint("NewApi")
-fun Fragment.checkWriteStoragePermission(): Boolean {
-    val writePermission = arrayOf(Manifest.permission.WRITE_EXTERNAL_STORAGE)
-    val activity = requireActivity()
-    return when {
-        Build.VERSION.SDK_INT < Build.VERSION_CODES.M -> true
-        requireContext().hasPermissions(writePermission) -> true
-        activity.requestPermissionsIsPossible(arrayOf(Manifest.permission.WRITE_EXTERNAL_STORAGE)) -> {
-            requestPermissions(writePermission, SyncUtils.REQUEST_WRITE_STORAGE_PERMISSION)
-            false
-        }
-        else -> {
-            MaterialAlertDialogBuilder(requireContext(), R.style.DialogStyle)
-                .setTitle(R.string.androidPermissionTitle)
-                .setMessage(R.string.allWritePermissionNeeded)
-                .setPositiveButton(R.string.buttonAuthorize) { _: DialogInterface?, _: Int ->
-                    activity.startAppSettingsConfig()
-                }
-                .show()
-            false
-        }
-    }
-}
-
-@SuppressLint("NewApi")
-fun Activity.checkWriteStoragePermission(): Boolean {
-    val writePermission = arrayOf(Manifest.permission.WRITE_EXTERNAL_STORAGE)
-    return when {
-        Build.VERSION.SDK_INT < Build.VERSION_CODES.M -> true
-        hasPermissions(writePermission) -> true
-        requestPermissionsIsPossible(arrayOf(Manifest.permission.WRITE_EXTERNAL_STORAGE)) -> {
-            requestPermissions(writePermission, SyncUtils.REQUEST_WRITE_STORAGE_PERMISSION)
-            false
-        }
-        else -> {
-            //TODO a better text message here: example, Additional permissions required to upload and download files.
-            MaterialAlertDialogBuilder(this, R.style.DialogStyle)
-                .setTitle(R.string.androidPermissionTitle)
-                .setMessage(R.string.allWritePermissionNeeded)
-                .setPositiveButton(R.string.buttonAuthorize) { _: DialogInterface?, _: Int ->
-                    startAppSettingsConfig()
-                }
-                .show()
-            false
-        }
-    }
-}
-
 fun View.hideKeyboard() {
     (context.getSystemService(Activity.INPUT_METHOD_SERVICE) as InputMethodManager)
         .hideSoftInputFromWindow(this.windowToken, 0)
